Move converted tasks to cogs directory
