+++ cogs/gru_nose.py
